Move ide-server
