+++ extensions/ql-vscode/src/language-support/ide-server.ts
@@ -1,8 +1,8 @@
 import { ProgressLocation, window } from "vscode";
 import { StreamInfo } from "vscode-languageclient/node";
-import { shouldDebugIdeServer, spawnServer } from "./cli";
-import { QueryServerConfig } from "./config";
-import { ideServerLogger } from "./common";
+import { shouldDebugIdeServer, spawnServer } from "../cli";
+import { QueryServerConfig } from "../config";
+import { ideServerLogger } from "../common";
 
 /**
  * Managing the language server for CodeQL.
